pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lostfound-backend"
version = "0.1.0"
description = "Scalable Arabic Facebook scraping + LLM intelligence pipeline for lost & found matching"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pgvector>=0.3.0",

    # AI / DSPy
    "dspy>=2.4.0",
    "google-generativeai>=0.7.0",

    # Scraping
    "apify-client>=1.7.0",
    "httpx>=0.27.0",

    # Config & validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",

    # Resilience
    "tenacity>=8.3.0",

    # Logging
    "structlog>=24.1.0",

    # Utilities
    "pyarabic>=0.6.15",  # Arabic text normalization
    "modal>=0.63.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
